test/udb_integration_test/requirements.txt
pyOCD==0.34.*
pexpect
pyserial
